+++ packages/sim-core/src/math/index.ts
@@ -1,0 +1,2 @@
+export * from './fixed'
+export * from './isqrt'